requirements.txt
kubernetes==28.1.0
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
colorama==0.4.6
click==8.1.7
pyyaml==6.0.1
rich==13.7.0
watchdog==3.0.0
regex==2023.10.3